feat: registra materia-prima

+++ start.sh
@@ -21,7 +21,7 @@ echo ''
 echo ''
 
 echo '  -- create .env --  \n'
-cp env.example .env
+[ -f .env ] || cp env.example .env
 echo '  -- build  -- \n'
 docker-compose up -d --build
